Moved ExcelDialect/Parser to plugins
