+++ frictionless/parsers/__init__.py
@@ -1,3 +1,2 @@
-from .excel import XlsxParser, XlsParser
 from .inline import InlineParser
 from .json import JsonParser, JsonlParser
